fix(demonsect): update name and url (#766)
* fix: update name and url

* fix: listing

* remove old demonsect source
