--- src/rust/madara/sources/demonsect/res/settings.json
@@ -1,16 +1,0 @@
-[
-	{
-		"type": "group",
-		"title": "Configurações",
-		"footer": "Nota: A maioria das fontes de Madara tem uma imagem de capa de alta resolução e uma imagem de capa de baixa resolução.\nUsar a imagem de alta resolução aumentará a quantidade de dados que precisa carregar e pode levar a um tempo de carregamento lento.",
-		"items": [
-			{
-				"type": "switch",
-				"key": "highres",
-				"title": "Capas de Alta Resolução",
-				"subtitle": "Para página inicial e de pesquisa",
-				"default": true
-			}
-		]
-	}
-]
