+++ src/rust/mangastream/sources/demonsect/res/filters.json
