+++ src/rust/mangastream/sources/demonsect/res/source.json
@@ -2,10 +2,10 @@
 	"info": {
 		"id": "pt-br.demonsect",
 		"lang": "pt-br",
-		"name": "Demon Sect",
-		"version": 2,
-		"url": "https://dsectcomics.org",
-		"nsfw": 2
+		"name": "Seita Celestial",
+		"version": 3,
+		"url": "https://seitacelestial.com",
+		"nsfw": 0
 	},
 	"listings": [
 		{
